Create component Settings :checkered_flag:
